requirements.txt
tqdm
matplotlib
torchvision
opencv-python==4.5.5.64
websockets
requests
pymysql
schedule
